requirements.txt
Flask>=2.0.0
python-dotenv>=0.19.0
qrcode[pil]>=7.0.0
psycopg2-binary>=2.9.0
requests>=2.25.0